updated ps3-goodrepos.sh

+++ ps3-goodrepos.sh
@@ -3,10 +3,15 @@
 # Script to remove (broken) repository configs installed with Fedora 12 (ppc) and replace them with good configs.
 # Also executes yum system update.
 #
-# RUN SCRIPT AS ROOT: sudo ./ps3-goodrepos.sh
+# RUN AS ROOT: sudo ./ps3-goodrepos.sh
 #
 # Note: Only the fedora and fedora-updates repos are enabled by default. Enable the others yourself if you wish - or don't.
 #
+if [ "$(id -u)" != 0 ]; then
+    printf "Please run as root.\n"
+    exit 1
+fi
+
 rm -fv /etc/pki/rpm-gpg/*
 cp -v rpm-gpg/RPM-GPG-KEY-fedora-12-primary /etc/pki/rpm-gpg/
 cp -v rpm-gpg/RPM-GPG-KEY-livna /etc/pki/rpm-gpg/
@@ -34,4 +39,6 @@ yum makecache
 yum -y update
 
 printf "\nREBOOT\n\n"
+
+exit 0
